Move Sync button to Inventory page
