+++ app/controllers/foreman_rh_cloud/react_controller.rb
@@ -7,8 +7,5 @@ module ForemanRhCloud
 
     def insights_cloud
     end
-
-    def inventory_sync
-    end
   end
 end
