--- app/views/foreman_rh_cloud/react/inventory_sync.html.erb
@@ -1,1 +1,0 @@
-<%= mount_react_component('InventorySync','#ForemanRhCloudReactRoot') %>
