+++ config/routes.rb
@@ -21,6 +21,5 @@ Rails.application.routes.draw do
   namespace :foreman_rh_cloud do
     get 'inventory_upload', to: 'react#inventory_upload'
     get 'insights_cloud', to: 'react#insights_cloud'
-    get 'inventory_sync', to: 'react#inventory_sync'
   end
 end
