+++ webpack/ForemanInventoryUpload/Components/PageHeader/components/SyncButton/SyncButtonHelpers.js
@@ -1,3 +1,3 @@
-import { foremanUrl } from '../ForemanRhCloudHelpers';
+import { foremanUrl } from '../../../../../ForemanRhCloudHelpers';
 
 export const inventorySyncUrl = path => foremanUrl(`/inventory_sync/${path}`);
