+++ webpack/ForemanInventoryUpload/Components/PageHeader/components/SyncButton/__tests__/InventorySyncHelpers.test.js
@@ -1,9 +1,9 @@
 import { testSelectorsSnapshotWithFixtures } from '@theforeman/test';
-import { inventorySyncUrl } from '../InventorySyncHelpers';
+import { inventorySyncUrl } from '../SyncButtonHelpers';
 
 const fixtures = {
   'should return insights cloud Url': () => inventorySyncUrl('test_path'),
 };
 
-describe('InventorySync helpers', () =>
+describe('SyncButton helpers', () =>
   testSelectorsSnapshotWithFixtures(fixtures));
