+++ webpack/ForemanInventoryUpload/Components/PageHeader/components/SyncButton/__tests__/SyncButton.test.js
@@ -1,0 +1,12 @@
+import { testComponentSnapshotsWithFixtures } from '@theforeman/test';
+
+import SyncButton from '../SyncButton';
+
+const fixtures = {
+  'render without Props': {},
+};
+
+describe('SyncButton', () => {
+  describe('rendering', () =>
+    testComponentSnapshotsWithFixtures(SyncButton, fixtures));
+});
