+++ webpack/ForemanInventoryUpload/Components/PageHeader/components/SyncButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './SyncButton';
