+++ webpack/ForemanInventoryUpload/Components/PageHeader/components/ToolbarButtons/ToolbarButtons.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import DocsButton from '../DocsButton';
 import HistoryButton from '../HistoryButton';
+import SyncButton from '../SyncButton';
 import './toolbarButtons.scss';
 
 const ToolbarButtons = () => (
   <div className="inventory_toolbar_buttons">
+    <SyncButton />
     <HistoryButton />
     <DocsButton />
   </div>
