+++ webpack/ForemanInventoryUpload/ForemanInventoryConstants.js
@@ -1,7 +1,9 @@
 import { translate as __ } from 'foremanReact/common/I18n';
 
-export const INVENTORY_PAGE_TITLE = __('Red Hat Inventory Uploads');
+export const INVENTORY_PAGE_TITLE = __('Red Hat Inventory');
 
 export const DOCS_BUTTON_TEXT = __(' Documentation');
 
 export const ACTIONS_HISTORY_BUTTON_TEXT = __(' Actions history');
+
+export const SYNC_BUTTON_TEXT = __(' Sync inventory status');
