+++ webpack/ForemanRhCloudPages.js
@@ -1,9 +1,7 @@
 import ForemanInventoryUpload from './ForemanInventoryUpload';
 import InsightsCloudSync from './InsightsCloudSync';
-import InventorySync from './InventorySync';
 
 export default [
   { name: 'ForemanInventoryUpload', type: ForemanInventoryUpload },
   { name: 'InsightsCloudSync', type: InsightsCloudSync },
-  { name: 'InventorySync', type: InventorySync },
 ];
