--- webpack/InventorySync/InventorySyncConstants.js
@@ -1,6 +1,0 @@
-import { translate as __ } from 'foremanReact/common/I18n';
-
-export const INVENTORY_SYNC_SUCCESS = 'INVENTORY_SYNC_SUCCESS';
-export const INVENTORY_SYNC_FAILURE = 'INVENTORY_SYNC_FAILURE';
-
-export const INSIGHTS_SYNC_PAGE_TITLE = __('Red Hat Inventory sync');
