--- webpack/InventorySync/InventorySyncReducers.js
@@ -1,3 +1,0 @@
-const reducers = {};
-
-export default reducers;
